Hierarchical Clustering Notebook
Hierarchical Clustering Using Scipy
In this notebook, we will be using Scipy to make dendrograms. The notebook gives the basic steps to realize a dendrogram from a numeric matrix. Let’s describe a few customizations that you can easily apply to your dendrogram based on the car dataset which contains different attributes like a model, mpg, gear, and other features related to cars.